requirements.txt
pillow
pystray
pywin32
keyboard
